feat: add webpack config, modules
